+++ admin/handlers/crud_comments.php
@@ -1,23 +1,25 @@
 <?php
-  // (UPDATE) 'APPROVE COMMENT'
-  if( isset( $_GET[ 'approve' ] ) ) {
-      $approve_id = $_GET[ 'approve' ];
-    
-      approveComment( $approve_id );
-  }
+// (UPDATE) 'APPROVE COMMENT'
+    if( isset( $_GET[ 'approve' ] ) ) {
+        $approve_id = escape( $_GET[ 'approve' ] );
+      
+        approveComment( $approve_id );
+    }
 
-  // (UPDATE) 'DENY COMMENT' 
-  if( isset( $_GET[ 'deny' ] ) ) {
-      $deny_id = $_GET[ 'deny' ];
-    
-      denyComment( $deny_id );
-  }
+// (UPDATE) 'DENY COMMENT' 
+    if( isset( $_GET[ 'deny' ] ) ) {
+        $deny_id = escape( $_GET[ 'deny' ] );
+      
+        denyComment( $deny_id );
+    }
 
-  // 'DELETE COMMENT'
-  if( isset( $_GET[ 'delete' ] ) ) {
-      $delete_id = $_GET[ 'delete' ];
-      $post_id   = $_GET[ 'post_id' ];
-    
-      deleteComment( $post_id, $delete_id ); 
-  }
+// 'DELETE COMMENT'
+    if( isset( $_GET[ 'delete' ] ) ) {
+        if( isset( $_SESSION['user_role'] ) && $_SESSION['user_role'] === 'admin' ) {
+            $delete_id = escape( $_GET[ 'delete' ] );
+            $post_id   = escape( $_GET[ 'post_id' ] );
+          
+            deleteComment( $post_id, $delete_id ); 
+        }
+    }
 ?>
